+++ packages/contracts/src/view/engine.ts
@@ -1,6 +1,7 @@
 'use strict'
 
 import { ViewConfig } from './config'
+import { HelperDelegate } from 'handlebars'
 
 export interface ViewEngine {
   /**
@@ -21,7 +22,23 @@ export interface ViewEngine {
   exists(view: string): Promise<boolean>
 
   /**
-   * Register a partial view with the given `name` and `content` to the handlebars engine.
+   * Register a partial view with the given `name` and `content` to the view engine.
    */
   registerPartial (name: string, content: string): this
+
+  /**
+   * Determine whether a partial view with the given `name` is registered.
+   */
+  hasPartial (name: string): boolean
+
+  /**
+   * Register a view helper with the given `name` and `content` to the view engine.
+   */
+  registerHelper (name: string, fn: HelperDelegate): this
+
+  /**
+   * Determine whether a view helper with the given `name` is registered.
+   */
+  hasHelper (name: string): boolean
+
 }