+++ packages/vite/src/index.ts
@@ -1,0 +1,10 @@
+'use strict'
+
+import { supercharge } from './plugin'
+export default supercharge
+
+export * from './contracts'
+export * from './inertia-helpers'
+export * from './plugin'
+export * from './vite'
+export * from './vite-service-provider'